+++ .eslintrc.js
@@ -1,37 +1,43 @@
 module.exports = {
   root: true,
+  parser: 'babel-eslint',
   parserOptions: {
-    ecmaVersion: 2017,
+    ecmaVersion: 2018,
     sourceType: 'module'
   },
+  globals: {
+    Highcharts: true,
+  },
   plugins: [
-    'ember',
-    'ember-best-practices'
+    'ember'
+  ],
+  extends: [
+    'eslint:recommended',
+    'plugin:ember/recommended'
   ],
   env: {
     browser: true
   },
-  globals: {
-    Highcharts: false,
-    define: false
-  },
   rules: {
-    'ember/avoid-leaking-state-in-ember-objects': 'off',
-    'ember-suave/no-direct-property-access': 'off'
   },
   overrides: [
     // node files
     {
       files: [
+        '.ember-cli.js',
+        '.eslintrc.js',
+        '.template-lintrc.js',
+        'ember-cli-build.js',
         'index.js',
         'testem.js',
-        'ember-cli-build.js',
+        'blueprints/*/index.js',
         'config/**/*.js',
         'tests/dummy/config/**/*.js'
       ],
       excludedFiles: [
-        'app/**',
         'addon/**',
+        'addon-test-support/**',
+        'app/**',
         'tests/dummy/app/**'
       ],
       parserOptions: {
@@ -44,7 +50,7 @@ module.exports = {
       },
       plugins: ['node'],
       rules: Object.assign({}, require('eslint-plugin-node').configs.recommended.rules, {
-        camelcase: 'off'
+        // add your custom rules and overrides for node files here
       })
     }
   ]