+++ app/components/high-charts.js
@@ -1,3 +1,1 @@
-import HighCharts from 'ember-highcharts/components/high-charts';
-
-export default HighCharts;
+export { default } from 'ember-highcharts/components/high-charts';