+++ config/ember-try.js
@@ -12,20 +12,28 @@ module.exports = function() {
       useYarn: true,
       scenarios: [
         {
-          name: 'ember-lts-3.0',
+          name: 'ember-lts-3.16',
           npm: {
             devDependencies: {
-              'ember-source': '~3.0.0'
-            }
-          }
+              'ember-source': '~3.16.0',
+            },
+          },
         },
         {
-          name: 'ember-lts-3.12',
+          name: 'ember-lts-3.20',
           npm: {
             devDependencies: {
-              'ember-source': '~3.12.0'
-            }
-          }
+              'ember-source': '~3.20.0',
+            },
+          },
+        },
+        {
+          name: 'ember-lts-3.24',
+          npm: {
+            devDependencies: {
+              'ember-source': '~3.24.0',
+            },
+          },
         },
         {
           name: 'ember-release',