+++ tests/dummy/app/components/chart-line-basic.js
@@ -1,8 +1,7 @@
-import Component from '@ember/component';
+import Component from '@glimmer/component';
 
-export default Component.extend({
-
-  chartOptions: {
+export default class LineBasic extends Component {
+  chartOptions = {
     chart: {
       type: 'line'
     },
@@ -30,9 +29,9 @@ export default Component.extend({
       verticalAlign: 'middle',
       borderWidth: 0
     }
-  },
+  }
 
-  chartData: [
+  chartData = [
     {
       name: 'Tokyo',
       data: [7.0, 6.9, 9.5, 14.5, 18.2, 21.5, 25.2, 26.5, 23.3, 18.3, 13.9, 9.6]
@@ -51,4 +50,4 @@ export default Component.extend({
     }
   ]
 
-});
+}