+++ CHANGELOG.md
@@ -1,7 +1,22 @@
 # Changelog
 
+## 3.0.0
+- Applied `ember init -b @ember/octane-addon-blueprint` for general upgrade goodness
+- Moved templates to co-location
+- Bumped `highcharts@^9.0.1`
+- Bumped `ember-copy@^2.0.1`
+- Bumped `devDeps` across the board
+- Implement Glimmer version of highcharts
+- Update tests; now with Angle Bracket Goodness™
+- Clean up `ember-try` config
+- Various lint fixes
+- Use `render-modifiers` bridge with `{{did-insert}}` and `{{did-update}}`
+- Export the build-options getter; move component test to util unit test
+- Drop node 10
+- Bring ember support to [LTS](https://emberjs.com/releases/lts/) with travis
+
 ## 2.2.0
-- Use ember-auto-import instead of ember-cli-node-assets
+- Use `ember-auto-import` instead of `ember-cli-node-assets`
 - Use peerDependencies and validate them
 
 ## 2.1.0